feat(m3-react,unstyled): reorganize exports add useStyles hook
- Export m3-react button module from m3-react index to centralize
  component access (export * as button from "./m3-react-button").
- Rename internal imports referring to button to use the new
  m3-react-button module (tests and bench files).
- Add useStyles hook (unstyled-use-styles.tsx) to provide a stable
  React useId-based style insertion API; move logic out of
  unstyled-cva and re-export from unstyled index.
- Rework unstyled package surface:
  - Replace single-string exports with an exports map so utilities can
    be imported as "unstyled/utilities".
  - Export unstyled-use-styles from package entry.
- Move utility helpers (mapValue, numberOrStringToString) into the
  proper utilities module and update imports accordingly (unstyled-cva
  now imports utilities from "./unstyled-utilities" and global design
  package imports "unstyled/utilities").
- Update storybook and app dependencies to include m3-react workspace
  package.
- Bump pnpm lock resolution for @napi-rs/wasm-runtime to 1.0.3.

These changes separate concerns between style utilities and the CVA
logic, make useStyles reusable across consumers, and standardize
package exports for better internal imports and workspace references.
