--- packages/m3-react/src/index.ts
@@ -1,1 +1,0 @@
-export {}
