+++ packages/unstyled/src/unstyled.ts
@@ -1,2 +1,3 @@
 export * from "./unstyled-cva"
 export * from "./unstyled-print"
+export * from "./unstyled-use-styles"
